feat: split OpenStack client module
Adding new auth methods makes single module too huge to be manageable.
Split out sync and async clients into separate modules, split out error.

This change breaks some internal interfaces that were public!
