+++ openstack_sdk/src/auth.rs
@@ -23,6 +23,8 @@ use tracing::{error, trace};
 
 use thiserror::Error;
 
+use futures::io::Error as IoError;
+
 use dialoguer::{Input, Password};
 
 use crate::api::identity::v3::auth::os_federation::{
@@ -32,10 +34,7 @@ use crate::api::identity::v3::auth::token::create as token_v3;
 use crate::api::identity::v3::auth::token::get as token_v3_info;
 use crate::api::RestEndpoint;
 use crate::config;
-use crate::types::identity::v3::{
-    self as types_v3, AuthReceiptResponse, AuthResponse, Domain, Project,
-};
-// TODO: complete adding error context through anyhow
+use crate::types::identity::v3::{AuthReceiptResponse, AuthResponse, Domain, Project};
 
 #[derive(Debug, Error)]
 #[non_exhaustive]
@@ -169,6 +168,25 @@ pub enum AuthError {
         #[from]
         source: token_v3_info::RequestBuilderError,
     },
+
+    #[error("WebSSO callback didn't return a token")]
+    WebSSONoToken,
+
+    #[error("WebSSO authentication failed")]
+    WebSSOFailed,
+
+    #[error("`IO` error: {}", source)]
+    IO {
+        #[from]
+        source: IoError,
+    },
+
+    #[error("`Join` error: {}", source)]
+    Join {
+        #[from]
+        source: tokio::task::JoinError,
+    },
+
     #[error(transparent)]
     Other(#[from] anyhow::Error), // source and Display delegate to anyhow::Error
 }
@@ -247,7 +265,7 @@ impl AuthType {
 }
 
 /// Fill Auth Request Identity with user password data
-pub fn fill_identity_using_password(
+fn fill_identity_using_password(
     identity_builder: &mut token_v3::IdentityBuilder<'_>,
     auth_data: &config::Auth,
     interactive: bool,
@@ -307,7 +325,7 @@ pub fn fill_identity_using_password(
 }
 
 /// Fill Auth Request Identity with user token
-pub fn fill_identity_using_token(
+fn fill_identity_using_token(
     identity_builder: &mut token_v3::IdentityBuilder<'_>,
     auth_data: &config::Auth,
     _interactive: bool,
@@ -321,7 +339,7 @@ pub fn fill_identity_using_token(
 }
 
 /// Fill Auth Request Identity with MFA passcode
-pub fn fill_identity_using_totp(
+fn fill_identity_using_totp(
     identity_builder: &mut token_v3::IdentityBuilder<'_>,
     auth_data: &config::Auth,
     interactive: bool,
@@ -399,7 +417,7 @@ fn process_auth_type(
     Ok(())
 }
 
-pub fn build_identity_data_from_config<'a>(
+pub(crate) fn build_identity_data_from_config<'a>(
     config: &config::CloudConfig,
     interactive: bool,
 ) -> Result<token_v3::Identity<'a>, AuthError> {
@@ -440,155 +458,11 @@ pub fn build_identity_data_from_config<'a>(
     Ok(identity_builder.build()?)
 }
 
-/// Build Auth `Identity` from existing `Auth` (use token)
-impl TryFrom<&AuthToken> for token_v3::Identity<'_> {
-    type Error = AuthError;
-
-    fn try_from(auth: &AuthToken) -> Result<Self, Self::Error> {
-        Ok(token_v3::IdentityBuilder::default()
-            .methods(Vec::from([token_v3::Methods::Token]))
-            .token(
-                token_v3::TokenBuilder::default()
-                    .id(auth.token.clone())
-                    .build()?,
-            )
-            .build()?)
-    }
-}
-
-/// Build Auth `Scope` data from `CloudConfig`
-impl TryFrom<&config::CloudConfig> for token_v3::Scope<'_> {
-    type Error = AuthError;
-    fn try_from(config: &config::CloudConfig) -> Result<Self, Self::Error> {
-        let auth = config.auth.clone().ok_or(AuthError::MissingAuthData)?;
-        let mut scope = token_v3::ScopeBuilder::default();
-        if auth.project_id.is_some() || auth.project_name.is_some() {
-            // Project scope
-            let mut project_scope = token_v3::ProjectBuilder::default();
-            if auth.project_domain_name.is_some() || auth.project_domain_id.is_some() {
-                let mut project_domain = token_v3::ProjectDomainBuilder::default();
-                if let Some(val) = auth.project_domain_id {
-                    project_domain.id(val);
-                }
-                if let Some(val) = auth.project_domain_name {
-                    project_domain.name(val);
-                }
-                project_scope.domain(project_domain.build()?);
-            };
-            if let Some(val) = auth.project_id {
-                project_scope.id(val);
-            }
-            if let Some(val) = auth.project_name {
-                project_scope.name(val);
-            }
-            scope.project(project_scope.build()?);
-        } else if auth.domain_id.is_some() || auth.domain_name.is_some() {
-            // Domain scope
-            let mut domain_scope = token_v3::ScopeDomainBuilder::default();
-            if let Some(val) = auth.domain_id {
-                domain_scope.id(val);
-            }
-            if let Some(val) = auth.domain_name {
-                domain_scope.name(val);
-            }
-            scope.domain(domain_scope.build()?);
-        } else {
-            return Err(AuthError::MissingScope);
-        }
-
-        Ok(scope.build()?)
-    }
-}
-
-/// Build Auth `Scope` data from existing `AuthorizationScope`
-impl TryFrom<&AuthorizationScope> for token_v3::Scope<'_> {
-    type Error = AuthError;
-    fn try_from(scope: &AuthorizationScope) -> Result<Self, Self::Error> {
-        let mut scope_builder = token_v3::ScopeBuilder::default();
-        match scope {
-            AuthorizationScope::Project(project) => {
-                let mut project_builder = token_v3::ProjectBuilder::default();
-                if let Some(val) = &project.id {
-                    project_builder.id(val.clone());
-                }
-                if let Some(val) = &project.name {
-                    project_builder.name(val.clone());
-                }
-                if let Some(domain) = &project.domain {
-                    let mut domain_builder = token_v3::ProjectDomainBuilder::default();
-                    if let Some(val) = &domain.id {
-                        domain_builder.id(val.clone());
-                    }
-                    if let Some(val) = &domain.name {
-                        domain_builder.name(val.clone());
-                    }
-                    project_builder.domain(domain_builder.build()?);
-                }
-                scope_builder.project(project_builder.build()?);
-            }
-            AuthorizationScope::Domain(domain) => {
-                let mut domain_builder = token_v3::ScopeDomainBuilder::default();
-                if let Some(val) = &domain.id {
-                    domain_builder.id(val.clone());
-                }
-                if let Some(val) = &domain.name {
-                    domain_builder.name(val.clone());
-                }
-                scope_builder.domain(domain_builder.build()?);
-            }
-            AuthorizationScope::Unscoped => {}
-        }
-        Ok(scope_builder.build()?)
-    }
-}
-
-/// Build `AuthorizationScope` data from `CloudConfig`
-impl TryFrom<&config::CloudConfig> for AuthorizationScope {
-    type Error = AuthError;
-    fn try_from(config: &config::CloudConfig) -> Result<Self, Self::Error> {
-        let auth = config.auth.clone().ok_or(AuthError::MissingAuthData)?;
-        if auth.project_id.is_some() || auth.project_name.is_some() {
-            // Project scope
-            Ok(AuthorizationScope::Project(Project {
-                id: auth.project_id.clone(),
-                name: auth.project_name.clone(),
-                domain: types_v3::get_domain(auth.project_domain_id, auth.project_domain_name),
-            }))
-        } else if auth.domain_id.is_some() || auth.domain_name.is_some() {
-            // Domain scope
-            Ok(AuthorizationScope::Domain(Domain {
-                id: auth.domain_id.clone(),
-                name: auth.domain_name.clone(),
-            }))
-        } else {
-            Ok(AuthorizationScope::Unscoped)
-        }
-    }
-}
-
-/// Build Auth request from `CloudConfig`
-impl TryFrom<&config::CloudConfig> for token_v3::Request<'_> {
-    type Error = AuthError;
-    fn try_from(config: &config::CloudConfig) -> Result<Self, Self::Error> {
-        let identity_data = build_identity_data_from_config(config, false)?;
-        let mut auth_request_data = token_v3::AuthBuilder::default();
-        auth_request_data.identity(identity_data);
-        if let Ok(scope_data) = token_v3::Scope::try_from(config) {
-            auth_request_data.scope(scope_data);
-        }
-
-        Ok(token_v3::RequestBuilder::default()
-            .auth(auth_request_data.build()?)
-            .build()?)
-    }
-}
-
 /// Build Auth request from `Identity` and `AuthScope`
 pub(crate) fn build_auth_request_with_identity_and_scope<'a>(
     auth: &token_v3::Identity<'a>,
     scope: &AuthorizationScope,
 ) -> Result<token_v3::Request<'a>, AuthError> {
-    //) -> Result<impl RestEndpoint + 'a, AuthError> {
     let mut auth_request_data = token_v3::AuthBuilder::default();
     auth_request_data.identity(auth.clone());
     if let Ok(scope_data) = token_v3::Scope::try_from(scope) {
@@ -664,6 +538,47 @@ pub(crate) fn build_auth_request_from_receipt<'a>(
             )]
             .iter()
             .cloned(),
+        )
+        .build()?)
+}
+
+/// Prepare Endpoint for the WebSSO authentication
+pub(crate) fn build_sso_auth_endpoint(
+    config: &config::CloudConfig,
+) -> Result<Cow<'static, str>, AuthError> {
+    if let Some(auth) = &config.auth {
+        if let Some(identity_provider) = &auth.identity_provider {
+            let mut ep = fed_idp_sso_get::RequestBuilder::default();
+            ep.idp_id(identity_provider);
+            if let Some(protocol) = &auth.protocol {
+                ep.protocol_id(protocol);
+            } else {
+                return Err(AuthError::AuthSsoMissingProtocol);
+            }
+            return Ok(ep.build()?.endpoint());
+        } else {
+            let mut ep = fed_sso_get::RequestBuilder::default();
+            if let Some(protocol) = &auth.protocol {
+                ep.protocol_id(protocol);
+            } else {
+                return Err(AuthError::AuthSsoMissingProtocol);
+            }
+            return Ok(ep.build()?.endpoint());
+        }
+    }
+    Err(AuthError::MissingAuthData)
+}
+
+pub(crate) fn build_token_info_endpoint<'a>(
+    subject_token: String,
+) -> Result<token_v3_info::Request<'a>, AuthError> {
+    Ok(token_v3_info::RequestBuilder::default()
+        .headers(
+            [(
+                Some(HeaderName::from_static("x-subject-token")),
+                HeaderValue::from_str(subject_token.as_str()).expect("Valid string"),
+            )]
+            .into_iter(),
         )
         .build()?)
 }
@@ -756,59 +671,6 @@ pub enum AuthorizationScope {
     Project(Project),
     Domain(Domain),
     Unscoped,
-}
-
-impl From<&AuthResponse> for AuthorizationScope {
-    fn from(auth: &AuthResponse) -> Self {
-        if let Some(project) = &auth.token.project {
-            Self::Project(project.clone())
-        } else if let Some(domain) = &auth.token.domain {
-            Self::Domain(domain.clone())
-        } else {
-            Self::Unscoped
-        }
-    }
-}
-
-/// Prepare Endpoint for the WebSSO authentication
-pub fn build_sso_auth_endpoint(
-    config: &config::CloudConfig,
-) -> Result<Cow<'static, str>, AuthError> {
-    if let Some(auth) = &config.auth {
-        if let Some(identity_provider) = &auth.identity_provider {
-            let mut ep = fed_idp_sso_get::RequestBuilder::default();
-            ep.idp_id(identity_provider);
-            if let Some(protocol) = &auth.protocol {
-                ep.protocol_id(protocol);
-            } else {
-                return Err(AuthError::AuthSsoMissingProtocol);
-            }
-            return Ok(ep.build()?.endpoint());
-        } else {
-            let mut ep = fed_sso_get::RequestBuilder::default();
-            if let Some(protocol) = &auth.protocol {
-                ep.protocol_id(protocol);
-            } else {
-                return Err(AuthError::AuthSsoMissingProtocol);
-            }
-            return Ok(ep.build()?.endpoint());
-        }
-    }
-    Err(AuthError::MissingAuthData)
-}
-
-pub fn build_token_info_endpoint<'a>(
-    subject_token: String,
-) -> Result<token_v3_info::Request<'a>, AuthError> {
-    Ok(token_v3_info::RequestBuilder::default()
-        .headers(
-            [(
-                Some(HeaderName::from_static("x-subject-token")),
-                HeaderValue::from_str(subject_token.as_str()).expect("Valid string"),
-            )]
-            .into_iter(),
-        )
-        .build()?)
 }
 
 #[cfg(test)]
